Alterando  o modelo de Inr tara double

+++ TaskPlannerMetrum/Model/ModelViews/vPM_MilestoneType.cs
@@ -7,11 +7,11 @@
         public int DepartmentID { get; set; }
         public string MilestonesValueID { get; set; }
         public int DisplacementServicesID { get; set; }
-        public float Hours { get; set; }
+        public double Hours { get; set; }
         public string ID { get; set; }
         public int FunctionID { get; set; }
         public string FunctionName { get; set; }
 
-        public int ValueHour {  get; set; } 
+        public double ValueHour {  get; set; } 
     }
 }
